:adhesive_bandage: (IOC): implemented inversify standarts
