+++ src/IOC/types.ts
@@ -1,0 +1,6 @@
+const TYPES = {
+	iUserRepository: Symbol.for("iUserRepository"),
+	iUserController: Symbol.for("iUserController")
+}
+
+export {TYPES}
