+++ src/controllers/userController.ts
@@ -1,12 +1,14 @@
 import { inject, injectable } from "inversify";
-import { UserRepository, iUserRepository } from "../repositories/UserRepository";
+import { UserRepository } from "../repositories/UserRepository";
 import { Request, Response } from 'express';
+import { iUserController, iUserRepository } from "../IOC/interfaces";
+import { TYPES } from "../IOC/types";
 
 @injectable()
 class UserController implements iUserController {
 
 	_userRepository;
-	constructor (@inject(UserRepository) private userRepository: iUserRepository) {
+	constructor (@inject(TYPES.iUserRepository) private userRepository: iUserRepository) {
 		this._userRepository = userRepository;
 	}
 
@@ -14,11 +16,5 @@ class UserController implements iUserController {
 		this._userRepository.getUsers(req, res)
 	}
 }
-
-export interface iUserController{
-	getAllUsers(req: Request, res: Response): void
-}
-
-
 
 export default UserController;
