+++ src/repositories/UserRepository.ts
@@ -1,6 +1,7 @@
 import { Request, Response } from 'express';
 import { PrismaClient } from '@prisma/client';
 import { injectable } from 'inversify';
+import { iUserRepository } from '../IOC/interfaces';
 
 @injectable()
 export class UserRepository implements iUserRepository {
@@ -10,8 +11,4 @@ export class UserRepository implements iUserRepository {
 		const users = await this.prisma.user.findMany()
 		res.json(users)
 	}
-}
-
-export interface iUserRepository {
-	getUsers(req: Request, res: Response): Promise<void>
 }
