+++ src/routes.ts
@@ -2,10 +2,11 @@ import 'reflect-metadata';
 import { Router } from 'express';
 import helloController from './controllers/helloController';
 import UserController from './controllers/userController';
-import { container } from './container';
+import { container } from './IOC/container';
+import { TYPES } from './IOC/types';
 
 const routes = Router();
-const userController = container.get<UserController>(UserController);
+const userController = container.get<UserController>(TYPES.iUserController);
 
 routes.get('/', helloController.getHelloMessage);
